requirements.txt
torch
transformers
ultralytics
opencv-python
Pillow
gTTS
IPython
numpy
